tests/wpt/web-platform-tests/app-history/navigate-event/signal-abort-window-stop-in-onnavigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = t.step_func_done(() => {
    let abort_signal;
    let onabort_called = false;
    let canceled_in_second_handler = false;
    appHistory.addEventListener("navigate", t.step_func(e => {
      abort_signal = e.signal;
      abort_signal.onabort = () => onabort_called = true;
      window.stop();
    }));
    appHistory.addEventListener("navigate", t.step_func(e => {
      canceled_in_second_handler = e.defaultPrevented;
    }));
    appHistory.navigate("?1");
    assert_true(abort_signal.aborted);
    assert_true(onabort_called);
    assert_true(canceled_in_second_handler);
  });
}, "window.stop() signals AppHistoryNavigateEvent.signal inside a navigate event handler");
</script>
